+++ src/lib.rs
@@ -83,7 +83,6 @@
 //!         image_data_len: 192,
 //!         channel_masks: None,
 //!         row_order: RowOrder::BottomUp,
-//!         color_table: None,
 //!     }
 //! );
 //!