+++ tests/chessboard-8px-1bit.rs
@@ -16,8 +16,12 @@ fn chessboard_8px_1bit() {
             image_data_len: 32,
             channel_masks: None,
             row_order: RowOrder::BottomUp,
-            color_table: Some(&[0, 0, 0, 0, 255, 255, 255, 255]),
         }
+    );
+
+    assert_eq!(
+        bmp.color_table(),
+        Some([0u8, 0, 0, 0, 255, 255, 255, 255].as_slice())
     );
 
     assert_eq!(bmp.image_data().len(), 94 - 62);